requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.8.2
jsonschema==4.23.0
requests==2.32.3

# OCR / PDF
pdfplumber==0.11.4
PyPDF2==3.0.1
pytesseract==0.3.13
Pillow==10.4.0
opencv-python-headless==4.10.0.84

# Google Sheets
gspread==6.1.4
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Slack
slack-bolt==1.20.1
slack-sdk==3.27.1

# Reliability
tenacity==9.0.0

# Testing
pytest==8.3.2
pytest-cov==5.0.0

# IBM Watsonx Orchestrate (ADK)
watsonx-orchestrate-adk==0.4.0 